+++ src/components/super/i-block/mods/index.ts
@@ -225,7 +225,7 @@ export default abstract class iBlockMods extends iBlockEvent {
 	 */
 	@hook('created')
 	protected initThemeModListener(): void {
-		if (this.r.theme == null) {
+		if (SSR || this.r.theme == null) {
 			return;
 		}