+++ CHANGELOG.md
@@ -11,10 +11,21 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v4.0.0-beta.?? (2024-??-??)
+## v4.0.0-beta.77 (2024-03-27)
+
+#### :rocket: New Feature
+
+* Implemented a new engine for translate keys `lang/engines`
+
+#### :house: Internal
+
+* Refactored the build ssr: all code has now been bundled into one file `build/graph`
+* Replaced the progress plugin's total time log with measure-loader `build/webpack/plugins/progress-plugin`
 
 #### :bug: Bug Fix
 
+* Resolved an issue with the progress plugin's view in relation to empty builds
+* Corrected the theme-manager to function properly in server-side rendering (ssr)
 * Added synchronous rendering of the first chunk using `v-for`. This is because SSR does not have access to the DOM API required for `vdom`. Therefore, we leverage Vue functionality to render the first chunk equally for SSR and CSR. `components/base/b-virtual-scroll-new`
 
 ## v4.0.0-beta.76 (2023-03-25)
@@ -178,12 +189,12 @@ All this will help avoid situations when for some reason the `IntersectionObserv
 
 * Added `preloadAmount` prop in b`-virtual-scroll-new` `components/base/b-virtual-scroll-new`
 
-#### :house:  Internal
+#### :house: Internal
 
 * Added re-export of modules from `tests/helpers`:
-  - Request
-  - RequestInterceptor
-  - Mock
+ - Request
+ - RequestInterceptor
+ - Mock
 
 * Removed bad import of async module from `tests/helpers/network/interceptor`